requirements.txt
# Web framework
fastapi>=0.115.0,<0.116.0
uvicorn[standard]>=0.30.0,<0.31.0

# Resume parsing
pdfplumber>=0.11.0,<0.12.0
docx2txt>=0.8

# OpenAI (new SDK)
openai>=1.45.0,<2.0.0

# Pinecone (new SDK, v6)
pinecone>=6.0.0,<7.0.0

# Environment variables
python-dotenv>=1.0.0,<2.0.0

# Optional - logging/typing
typing-extensions>=4.12.0,<5.0.0

# Optional frontend (only if using Streamlit)
streamlit>=1.38.0,<2.0.0

python-multipart

uvicorn